migrate io.rs to services
